+++ back-end/src/models/student.ts
@@ -1,0 +1,28 @@
+import { Entity, Column, PrimaryColumn } from "typeorm";
+
+//define student table
+@Entity()
+export class Student {
+    @PrimaryColumn()
+    id!: number
+
+    @Column({
+        length: 100,
+    })
+    name!: string
+
+    @Column()
+    gender!: string
+
+    @Column()
+    address!: string
+
+    @Column()
+    mobile!: string
+
+    @Column()
+    birthday!: string
+
+    @Column()
+    age!: number
+}
