+++ back-end/src/routes/index.ts
@@ -1,14 +1,10 @@
 import express from 'express';
-import addStudent from '../controllers/addStudent';
-import getStudents from '../controllers/getStudents';
-import updateStudent from '../controllers/updateStudent';
-import deleteStudent from '../controllers/deleteStudent';
-
+import {addStudent,getStudents, deleteStudent, updateStudent} from '../controllers/student';
 //add routers from controllers
 const router = express.Router();
 router.post('/add-student',addStudent)
 router.get('/get-students', getStudents);
 router.post('/update-student', updateStudent);
-router.post('/remove-student', deleteStudent)
+router.post('/remove-student', deleteStudent);
 
 export default router;
